requirements.txt
ultralytics
uv
opencv-python
numpy
pandas
scikit-learn
roboflow
torchvision